requirements.txt
fastapi
uvicorn[standard]
selenium
flet
python-dotenv
sqlalchemy
psycopg2-binary
pydantic
